refactor: centralize exit codes and endpoint file constants
