+++ src/cli/core/constants.ts
@@ -1,4 +1,3 @@
-export const END_POINT_FILE_NAMES = ["page.tsx", "route.ts"] as const;
 export const QUERY_TYPES = ["Query", "OptionalQuery"] as const;
 
 export const INDENT = "  ";
