--- src/cli/core/types.ts
@@ -1,3 +1,0 @@
-import type { END_POINT_FILE_NAMES } from "./constants";
-
-export type EndPointFileNames = (typeof END_POINT_FILE_NAMES)[number];
